2023-08-09 1.追加oracle数据源

+++ datavines-engine/datavines-engine-plugins/datavines-engine-local/datavines-engine-local-transform/src/main/java/io/datavines/engine/local/transform/sql/InvalidateItemsExecutor.java
@@ -16,6 +16,7 @@
  */
 package io.datavines.engine.local.transform.sql;
 
+//import com.baomidou.mybatisplus.annotation.DbType;
 import io.datavines.common.config.Config;
 import io.datavines.engine.local.api.entity.ResultList;
 import lombok.extern.slf4j.Slf4j;
@@ -35,8 +36,15 @@ public class InvalidateItemsExecutor implements ITransformExecutor {
         String sql = config.getString(SQL);
 
         Statement statement = connection.createStatement();
-        statement.execute("DROP VIEW IF EXISTS " + outputTable);
-        statement.execute("CREATE VIEW " + outputTable + " AS " + sql);
+
+        if("oracle".equalsIgnoreCase(connection.getMetaData().getDatabaseProductName())) {
+            statement.execute("CREATE OR REPLACE VIEW " + outputTable + " AS " + sql);
+        } else {
+
+            statement.execute("DROP VIEW IF EXISTS " + outputTable);
+            statement.execute("CREATE VIEW " + outputTable + " AS " + sql);
+        }
+
         statement.close();
         return null;
     }
